requirements.txt
pycryptodome
lz4
